feature/monorepo-rename-HttpKernel-to-Kernel-in-tests
