+++ packages/EasyApiToken/tests/Unit/AbstractSymfonyTestCase.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace EonX\EasyApiToken\Tests\Unit;
 
-use EonX\EasyApiToken\Tests\Stub\HttpKernel\KernelStub;
+use EonX\EasyApiToken\Tests\Stub\Kernel\KernelStub;
 use Symfony\Component\HttpKernel\KernelInterface;
 
 abstract class AbstractSymfonyTestCase extends AbstractUnitTestCase
