+++ packages/EasyBugsnag/tests/EonX/EasyBugsnag/Tests/Stub/Kernel/KernelStub.php
@@ -1,19 +1,22 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyRequestId\Tests\Stub\HttpKernel;
+namespace EonX\EasyBugsnag\Tests\Stub\Kernel;
 
-use EonX\EasyRandom\Bundle\EasyRandomBundle;
-use EonX\EasyRequestId\Bundle\EasyRequestIdBundle;
+use EonX\EasyBugsnag\Bundle\EasyBugsnagBundle;
+use EonX\EasyUtils\Bundle\EasyUtilsBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpKernel\Kernel;
 
 final class KernelStub extends Kernel implements CompilerPassInterface
 {
+    private static ?Request $request = null;
+
     /**
      * @var string[]
      */
@@ -29,16 +32,35 @@ final class KernelStub extends Kernel implements CompilerPassInterface
         parent::__construct('test', true);
     }
 
+    public static function createRequestStack(): RequestStack
+    {
+        $requestStack = new RequestStack();
+
+        if (self::$request !== null) {
+            $requestStack->push(self::$request);
+        }
+
+        return $requestStack;
+    }
+
+    public static function setRequest(Request $request): void
+    {
+        self::$request = $request;
+    }
+
     public function process(ContainerBuilder $container): void
     {
-        $container->setDefinition(RequestStack::class, new Definition(RequestStack::class));
+        $requestStackDef = new Definition(RequestStack::class);
+        $requestStackDef->setFactory([self::class, 'createRequestStack']);
+
+        $container->setDefinition(RequestStack::class, $requestStackDef);
 
         foreach ($container->getAliases() as $alias) {
             $alias->setPublic(true);
         }
 
-        foreach ($container->getDefinitions() as $definition) {
-            $definition->setPublic(true);
+        foreach ($container->getDefinitions() as $def) {
+            $def->setPublic(true);
         }
     }
 
@@ -47,10 +69,13 @@ final class KernelStub extends Kernel implements CompilerPassInterface
      */
     public function registerBundles(): iterable
     {
-        yield new EasyRandomBundle();
-        yield new EasyRequestIdBundle();
+        yield new EasyBugsnagBundle();
+        yield new EasyUtilsBundle();
     }
 
+    /**
+     * @throws \Exception
+     */
     public function registerContainerConfiguration(LoaderInterface $loader): void
     {
         foreach ($this->configs as $config) {
