+++ packages/EasyEncryption/tests/EonX/EasyEncryption/Tests/Stub/Kernel/KernelStub.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyRandom\Tests\Stub\HttpKernel;
+namespace EonX\EasyEncryption\Tests\Stub\Kernel;
 
-use EonX\EasyRandom\Bundle\EasyRandomBundle;
+use EonX\EasyEncryption\Bundle\EasyEncryptionBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -19,17 +19,21 @@ final class KernelStub extends Kernel implements CompilerPassInterface
     /**
      * @param string[]|null $configs
      */
-    public function __construct(string $environment, bool $debug, ?array $configs = null)
+    public function __construct(?array $configs = null)
     {
         $this->configs = $configs ?? [];
 
-        parent::__construct($environment, $debug);
+        parent::__construct('test', true);
     }
 
     public function process(ContainerBuilder $container): void
     {
-        foreach ($container->getDefinitions() as $def) {
-            $def->setPublic(true);
+        foreach ($container->getAliases() as $alias) {
+            $alias->setPublic(true);
+        }
+
+        foreach ($container->getDefinitions() as $definition) {
+            $definition->setPublic(true);
         }
     }
 
@@ -38,9 +42,12 @@ final class KernelStub extends Kernel implements CompilerPassInterface
      */
     public function registerBundles(): iterable
     {
-        yield new EasyRandomBundle();
+        yield new EasyEncryptionBundle();
     }
 
+    /**
+     * @throws \Exception
+     */
     public function registerContainerConfiguration(LoaderInterface $loader): void
     {
         foreach ($this->configs as $config) {
