+++ packages/EasyEncryption/tests/Unit/bundle/AbstractSymfonyTestCase.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace EonX\EasyEncryption\Tests\Unit\Bundle;
 
-use EonX\EasyEncryption\Tests\Stub\HttpKernel\KernelStub;
+use EonX\EasyEncryption\Tests\Stub\Kernel\KernelStub;
 use EonX\EasyEncryption\Tests\Unit\AbstractUnitTestCase;
 use Symfony\Component\HttpKernel\KernelInterface;
