+++ packages/EasyLock/tests/Stub/Kernel/KernelStub.php
@@ -1,13 +1,15 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyLogging\Tests\Stub\HttpKernel;
+namespace EonX\EasyLock\Tests\Stub\Kernel;
 
-use EonX\EasyLogging\Bundle\EasyLoggingBundle;
-use EonX\EasyUtils\Bundle\EasyUtilsBundle;
+use EonX\EasyLock\Bundle\EasyLockBundle;
+use Psr\Log\LoggerInterface;
+use Psr\Log\NullLogger;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\HttpKernel\Kernel;
 
 final class KernelStub extends Kernel implements CompilerPassInterface
@@ -29,6 +31,8 @@ final class KernelStub extends Kernel implements CompilerPassInterface
 
     public function process(ContainerBuilder $container): void
     {
+        $container->setDefinition(LoggerInterface::class, new Definition(NullLogger::class));
+
         foreach ($container->getAliases() as $alias) {
             $alias->setPublic(true);
         }
@@ -43,13 +47,9 @@ final class KernelStub extends Kernel implements CompilerPassInterface
      */
     public function registerBundles(): iterable
     {
-        yield new EasyLoggingBundle();
-        yield new EasyUtilsBundle();
+        yield new EasyLockBundle();
     }
 
-    /**
-     * @throws \Exception
-     */
     public function registerContainerConfiguration(LoaderInterface $loader): void
     {
         foreach ($this->configs as $config) {
