+++ packages/EasyNotification/tests/EonX/EasyNotification/Tests/Stub/Kernel/KernelStub.php
@@ -1,22 +1,16 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyBugsnag\Tests\Stub\HttpKernel;
+namespace EonX\EasyNotification\Tests\Stub\Kernel;
 
-use EonX\EasyBugsnag\Bundle\EasyBugsnagBundle;
-use EonX\EasyUtils\Bundle\EasyUtilsBundle;
+use EonX\EasyNotification\Bundle\EasyNotificationBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Definition;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpKernel\Kernel;
 
 final class KernelStub extends Kernel implements CompilerPassInterface
 {
-    private static ?Request $request = null;
-
     /**
      * @var string[]
      */
@@ -32,29 +26,8 @@ final class KernelStub extends Kernel implements CompilerPassInterface
         parent::__construct('test', true);
     }
 
-    public static function createRequestStack(): RequestStack
-    {
-        $requestStack = new RequestStack();
-
-        if (self::$request !== null) {
-            $requestStack->push(self::$request);
-        }
-
-        return $requestStack;
-    }
-
-    public static function setRequest(Request $request): void
-    {
-        self::$request = $request;
-    }
-
     public function process(ContainerBuilder $container): void
     {
-        $requestStackDef = new Definition(RequestStack::class);
-        $requestStackDef->setFactory([self::class, 'createRequestStack']);
-
-        $container->setDefinition(RequestStack::class, $requestStackDef);
-
         foreach ($container->getAliases() as $alias) {
             $alias->setPublic(true);
         }
@@ -69,13 +42,9 @@ final class KernelStub extends Kernel implements CompilerPassInterface
      */
     public function registerBundles(): iterable
     {
-        yield new EasyBugsnagBundle();
-        yield new EasyUtilsBundle();
+        yield new EasyNotificationBundle();
     }
 
-    /**
-     * @throws \Exception
-     */
     public function registerContainerConfiguration(LoaderInterface $loader): void
     {
         foreach ($this->configs as $config) {
