+++ packages/EasyRandom/tests/Stub/Kernel/KernelStub.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyNotification\Tests\Stub\HttpKernel;
+namespace EonX\EasyRandom\Tests\Stub\Kernel;
 
-use EonX\EasyNotification\Bundle\EasyNotificationBundle;
+use EonX\EasyRandom\Bundle\EasyRandomBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -19,19 +19,15 @@ final class KernelStub extends Kernel implements CompilerPassInterface
     /**
      * @param string[]|null $configs
      */
-    public function __construct(?array $configs = null)
+    public function __construct(string $environment, bool $debug, ?array $configs = null)
     {
         $this->configs = $configs ?? [];
 
-        parent::__construct('test', true);
+        parent::__construct($environment, $debug);
     }
 
     public function process(ContainerBuilder $container): void
     {
-        foreach ($container->getAliases() as $alias) {
-            $alias->setPublic(true);
-        }
-
         foreach ($container->getDefinitions() as $def) {
             $def->setPublic(true);
         }
@@ -42,7 +38,7 @@ final class KernelStub extends Kernel implements CompilerPassInterface
      */
     public function registerBundles(): iterable
     {
-        yield new EasyNotificationBundle();
+        yield new EasyRandomBundle();
     }
 
     public function registerContainerConfiguration(LoaderInterface $loader): void
